FEM: new solver equations, use small character for equations objects
