+++ src/Mod/Fem/App/CMakeLists.txt
@@ -127,7 +127,7 @@ SET(FemObjectsScripts_SRCS
 SET(FemSolver_SRCS
     FemSolver/__init__.py
     FemSolver/solverbase.py
-    FemSolver/EquationBase.py
+    FemSolver/equationbase.py
 )
 
 SET(FemElmer_SRCS
@@ -139,14 +139,14 @@ SET(FemElmer_SRCS
 )
 
 SET(FemEquationsElmer_SRCS
-    FemSolver/elmer/Equations/__init__.py
-    FemSolver/elmer/Equations/Base.py
-    FemSolver/elmer/Equations/Linear.py
-    FemSolver/elmer/Equations/Nonlinear.py
-    FemSolver/elmer/Equations/Elasticity.py
-    FemSolver/elmer/Equations/Electrostatic.py
-    FemSolver/elmer/Equations/Heat.py
-    FemSolver/elmer/Equations/Flow.py
+    FemSolver/elmer/equations/__init__.py
+    FemSolver/elmer/equations/base.py
+    FemSolver/elmer/equations/linear.py
+    FemSolver/elmer/equations/nonlinear.py
+    FemSolver/elmer/equations/elasticity.py
+    FemSolver/elmer/equations/electrostatic.py
+    FemSolver/elmer/equations/heat.py
+    FemSolver/elmer/equations/flow.py
 )
 
 SET(FemCalculix_SRCS
